backend/requirements.txt
langgraph==1.1.6
langchain==1.2.15
langchain-community==0.4.1
langchain-google-genai==4.2.1
duckduckgo-search==8.1.1
ddgs==9.13.0
python-dotenv==1.2.2
fastapi==0.135.3
uvicorn==0.44.0
